Fix after review comments
